refactor: update authentication model imports
