+++ data/src/main/java/com/london/data/remote/model/authentication/SessionResponse.kt
@@ -1,4 +1,4 @@
-package com.london.data.remote.model.authentication.model
+package com.london.data.remote.model.authentication
 
 import kotlinx.serialization.SerialName
 import kotlinx.serialization.Serializable
