+++ data/src/main/java/com/london/data/remote/source/authentication/AuthenticationRemoteDataSource.kt
@@ -1,8 +1,8 @@
 package com.london.data.remote.source.authentication
-import com.london.data.remote.model.authentication.model.GuestSessionResponse
-import com.london.data.remote.model.authentication.model.RequestTokenResponse
-import com.london.data.remote.model.authentication.model.SessionResponse
-import com.london.data.remote.model.authentication.model.DeleteSessionResponse
+import com.london.data.remote.model.authentication.DeleteSessionResponse
+import com.london.data.remote.model.authentication.GuestSessionResponse
+import com.london.data.remote.model.authentication.RequestTokenResponse
+import com.london.data.remote.model.authentication.SessionResponse
 
 interface AuthenticationRemoteDataSource {
     suspend fun createRequestToken(): Result<RequestTokenResponse>
